refactor: move generateBuilderPageStaticParams
