+++ src/app/(builder)/[countryCode]/content/[[...page]]/page.tsx
@@ -4,8 +4,8 @@ import Script from 'next/script'
 import { BuilderPageLayout, RenderBuilderContent } from '@/components/app/builder'
 import { PageProps } from '@/types'
 import { BuilderPageModelIdentifiers } from '@/utils/server/builder/models/page/constants'
+import { generateBuilderPageStaticParams } from '@/utils/server/builder/models/page/generateBuilderPageStatisParams'
 import { getPageContent, getPageDetails } from '@/utils/server/builder/models/page/utils'
-import { generateBuilderPageStaticParams } from '@/utils/server/builder/models/page/utils/generateBuilderPageStatisParams'
 import { generateMetadataDetails } from '@/utils/server/metadataUtils'
 
 export const dynamic = 'error'
