+++ src/app/(builder)/[countryCode]/press/[...page]/page.tsx
@@ -3,8 +3,8 @@ import { Metadata } from 'next'
 import { BuilderPageLayout, RenderBuilderContent } from '@/components/app/builder'
 import { PageProps } from '@/types'
 import { BuilderPageModelIdentifiers } from '@/utils/server/builder/models/page/constants'
+import { generateBuilderPageStaticParams } from '@/utils/server/builder/models/page/generateBuilderPageStatisParams'
 import { getPageContent, getPageDetails } from '@/utils/server/builder/models/page/utils'
-import { generateBuilderPageStaticParams } from '@/utils/server/builder/models/page/utils/generateBuilderPageStatisParams'
 import { generateMetadataDetails } from '@/utils/server/metadataUtils'
 
 export const dynamic = 'error'
